pyproject.toml
[tool.poetry]
name = "weatherapp"
version = "0.1.0"
description = ""
authors = ["ericcarry88828 <ericcarry88828@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.1.0"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
redis = "^5.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
